+++ pkgs/dart_mcp_server/lib/src/utils/analytics.dart
@@ -164,6 +164,7 @@ enum CallToolFailureReason {
   noRootGiven,
   noRootsSet,
   noSuchCommand,
+  nonZeroExitCode,
   mustSpecifyDtdUri,
   processException,
   timeout,